move to dev_utils as that seems more suitable location
